+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+- feat: add `onFocusChange` callback to `SelectionGroup`
+- feat: suppress `WidgetState.selected` on all items while group has focus
+
 ## 0.0.3
 
 - fix: update LICENSE to MIT